requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipykernel==6.29.5
ipython==8.14.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.1.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
